MPD.listAlbums: supply scope modifier.
In practice it works just fine without it, but since
http://mpd.wikia.com/wiki/MusicPlayerDaemonCommands#list says its required
this behaviour might be a bug. I think the documentation is just outdated,
as 'list' seems to be rather clever (e.g. it does the right thing even
when there are files with fields other than 'artist' that match the query
string). Yet I find it best to comply with the official docs.

darcs-hash:20070807124257-ea16c-f28617281152dec22a21f11f85a983b74d5d13db.gz