Prim.splitAck: fix.
Turns out that the ACK parsing was wrong.
The number after '@' is the offset of the command that caused the error in a
command list.
splitAck has been updated to return the actual error code, without the offset,
along with the name of the last run command and the error message from MPD.

darcs-hash:20070901210936-ea16c-bcccbff9796e08f96bd777d5fcac90c2a9ebd751.gz