mpdReceive: read lines from the socket lazily
This significantly decreases memory usage, BUT there is a slight problem.
With that patch data from the socket is not read until parse phase. Thus,
if parsing fails in the middle with 'Unexpected response', there will be
data left in the socket. This can be either resolved by catching Unexpected
exception in parsers and forcing library to read what's left in the socket,
have generic field for 'unknown' values or simply discard them. The latter
seems to be most reasonable.