MPD.Core: make mpdSend more robust to MPD connection errors
Previously, mpdSend would not catch exceptions caused by attempts to
write to an invalid connection handle.
This patch simplifies the error handling logic and makes sure to
handle failed writes in the same way as failed attempts to read the
response (i.e., invalidate the session handle and re-execute the
request, triggering an attempt to re-connect).