Support latest GHC (and latest stackage nightly)
This PR should be properly reviewed as I have only made sure that it
compiles fine and the test suites are passed.

Summary of the changes:
* Only support latest 4 major GHC's. With xmonad also doing the same,
I made the same decision here.
* In recent version of network (>= 3.0), the connectTo function isn't
present. So, safeConnect function has been changed accordingly. (This
is the part that need review from the person actually using the library).

Related to: https://github.com/xmonad/xmonad-extras/issues/24